TST: Re-set-up nose test coverage and CodeCov #156
Basically replicates what I did a while back in
74d0f46b4ec688726b455baa05ccfd7ed91c6f87 and
f84fa788e992c09e42e269f24cc716680f157fb1 (with the slight change
that now the coverage package is installed in setup.py's test
requirements, which is a pretty nifty way to handle this).

Partial progress on #156; still need to finagle things so that we
can get test coverage from the JS tests.