begin work on tidying up CSS: clean up menu CSS
based on the #207 PR work.

This is a good start, but there's a funky space below the node ID.
not sure what the cause is -- I think default browser styling stuff
is probably a factor, but it can't be the only reason (since this
change isn't present before these CSS trimming fixes).