ENH: draw symmetric + width-meaningful clades #342
This actually looks nice! However, we still would need to add in
logic that detects if the midY / midAngle points are "covered"
by the parent of the root's internal node, and if not connecting
that line I guess? Or altering logic in getCoords() / draw thick
lines / etc. to do that (via registering midY as the new y coord
of the root node i suppose, idk...?)

ANYWAY this shouldn't block #380 but it's cool to have in as proof
of concept